requirements.txt
Flask
Flask-Bcrypt
transformers
torch
pandas
google-api-python-client>=2.84.0
google-generativeai>=0.7.1
langdetect
accelerate
matplotlib
datasets
python-dotenv
scikit-learn
pydantic
pytz
